Software/requirements.txt
numpy==2.1.1
pyserial==3.5
